+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [0.3.5] - 2019-09-13
+### Added
 - HelpLink for displaying a button to open a help page.
+- `parent_field` attribute for fields. Set this to the name of an encompasing Hash (JSONB) column.
 ### Changed
 - Styling of fold_ups improved.
 - Expand/Collapse can be called within a Column.
-### Fixed
+- `invisible` property of fields can be set and if true, the field will not be rendered.
 
 ## [0.3.4] - 2019-09-13
 ### Added